+++ bos/hal/inc/b_hal_flash.h
@@ -56,7 +56,7 @@ extern "C" {
 int      bMcuFlashInit(void);
 int      bMcuFlashUnlock(void);
 int      bMcuFlashLock(void);
-int      bMcuFlashErase(uint32_t raddr, uint8_t pages);
+int      bMcuFlashErase(uint32_t raddr, uint32_t pages);
 int      bMcuFlashWrite(uint32_t raddr, const uint8_t *pbuf, uint32_t len);
 int      bMcuFlashRead(uint32_t raddr, uint8_t *pbuf, uint32_t len);
 uint32_t bMcuFlashSectorSize(void);
@@ -66,7 +66,7 @@ uint32_t bMcuFlashChipSize(void);
 int      bHalFlashInit(void);
 int      bHalFlashUnlock(void);
 int      bHalFlashLock(void);
-int      bHalFlashErase(uint32_t raddr, uint8_t pages);
+int      bHalFlashErase(uint32_t raddr, uint32_t pages);
 int      bHalFlashWrite(uint32_t raddr, const uint8_t *pbuf, uint32_t len);
 int      bHalFlashRead(uint32_t raddr, uint8_t *pbuf, uint32_t len);
 uint32_t bHalFlashSectorSize(void);
